video embeds

+++ supabase/add_video_timestamps.sql
@@ -1,7 +1,7 @@
 alter table tricks
-  add column video_start integer,
-  add column video_end   integer;
+  add column video_start smallint,
+  add column video_end   smallint;
 
 alter table user_tricks
-  add column video_start integer,
-  add column video_end   integer;
+  add column video_start smallint,
+  add column video_end   smallint;
